+++ test/development/basic/next-dynamic/next-dynamic-babel-src-dir.test.ts
@@ -1,0 +1,3 @@
+process.env.TEST_BABEL = '1'
+process.env.TEST_SRC_DIR = '1'
+require('./next-dynamic.test')
