+++ test/development/basic/next-dynamic/next-dynamic-babel.test.ts
@@ -1,0 +1,2 @@
+process.env.TEST_BABEL = '1'
+require('./next-dynamic.test')
