+++ test/development/basic/next-dynamic/next-dynamic-base-path.test.ts
@@ -1,0 +1,2 @@
+process.env.TEST_BASE_PATH = '/docs'
+require('./next-dynamic.test')
