+++ test/development/basic/next-dynamic/next-dynamic-custom-document.test.ts
@@ -1,0 +1,2 @@
+process.env.TEST_CUSTOMIZED_DOCUMENT = '1'
+require('./next-dynamic.test')
